Started to make steps independent of other steps.
Im sorry about how big the commit is. I still am running into issue with the model being called to early, which is due to the import statements when segmenting the code into individual files. I might keep of the Util functions in one file for the time being until I can resolve the imports